Module firewood::db

Structs

DB

Firewood database handle.

DBConfig

Database configuration.

DBRev

Some readable version of the DB.

DBRevConfig

Config for accessing a version of the DB.

DiskBufferConfig

Config for the disk buffer.

Revision

Lock protected handle to a readable version of the DB.

WALConfig

WriteBatch

An atomic batch of changes made to the DB. Each operation on a WriteBatch will move itself because when an error occurs, the write batch will be automaticlaly aborted so that the DB remains clean.

Enums

DBError